+++ lib/infrastructure/screens/home/home_screen.dart
@@ -5,9 +5,8 @@ import 'package:gymnastic_center/presentation/widgets/course/course_carrusel.dar
 import 'package:gymnastic_center/infrastructure/services/categories/CategoryService.dart';
 import 'package:gymnastic_center/presentation/widgets/home/category_carousel.dart';
 import 'package:gymnastic_center/infrastructure/services/blogs/BlogsService.dart';
-import 'package:gymnastic_center/presentation/widgets/home/BlogList.dart';
+import 'package:gymnastic_center/presentation/widgets/home/blog_list.dart';
 import 'package:gymnastic_center/presentation/widgets/home/home_app_bar.dart';
-import 'package:gymnastic_center/presentation/widgets/home/category_carousel.dart';
 
 class HomeScreen extends StatelessWidget {
   const HomeScreen({super.key});
@@ -15,24 +14,22 @@ class HomeScreen extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     final popularCourses = context.watch<CourseBloc>().state.popularCourses;
-    
+
     return Scaffold(
       appBar: const PreferredSize(
           preferredSize: Size(double.infinity, 230), child: HomeAppBar()),
-body: Column(
+      body: Column(
         children: [
           CategoryCarousel(categoryRepository: CategoryService()),
-          SizedBox(height: 20), 
-           Expanded(
-    child: BlogList(blogRepository: BlogsService()),
-  ),
+          const SizedBox(height: 20),
+          Expanded(child: BlogList(blogRepository: BlogsService())),
           Padding(
-            padding: const EdgeInsets.symmetric(horizontal: 15),
-            child: CourseCarrusel(
-              courses: popularCourses,
-            )),
+              padding: const EdgeInsets.symmetric(horizontal: 15),
+              child: CourseCarrusel(
+                courses: popularCourses,
+              )),
         ],
       ),
-              );
-            }
-          }
+    );
+  }
+}
