+++ lib/presentation/widgets/home/blog_list.dart
@@ -1,0 +1,133 @@
+import 'package:gymnastic_center/domain/blog/repository/IBlogRepository.dart';
+import 'package:gymnastic_center/presentation/widgets/blog/Blog_Detail.dart';
+import 'package:gymnastic_center/application/models/blog.dart';
+import 'package:flutter/material.dart';
+
+class BlogList extends StatelessWidget {
+  final IBlogRepository blogRepository;
+
+  const BlogList({super.key, required this.blogRepository});
+
+  @override
+  Widget build(BuildContext context) {
+    return FutureBuilder<List<Blog>>(
+      future: blogRepository.getAllBlogs(),
+      builder: (context, snapshot) {
+        if (snapshot.connectionState == ConnectionState.waiting) {
+          return Container();
+        } else if (snapshot.hasError) {
+          return Text('Error: ${snapshot.error}');
+        } else if (snapshot.hasData) {
+          List<Blog> blogs = snapshot.data!;
+          return Column(
+            children: [
+              const Padding(
+                padding: EdgeInsets.all(10.0),
+                child: Align(
+                  alignment: Alignment.centerLeft,
+                  child: Text(
+                    'Our Latest Blogs',
+                    style: TextStyle(
+                      color: Colors.black,
+                      fontSize: 18,
+                      fontWeight: FontWeight.bold,
+                    ),
+                  ),
+                ),
+              ),
+              Flexible(
+                child: GridView.builder(
+                  gridDelegate: const SliverGridDelegateWithFixedCrossAxisCount(
+                    crossAxisCount: 2,
+                    childAspectRatio: 1,
+                  ),
+                  itemCount: blogs.length,
+                  itemBuilder: (context, index) {
+                    return SizedBox(
+                      width: 250,
+                      height: 209,
+                      child: Padding(
+                        padding: const EdgeInsets.all(10.0),
+                        child: Stack(
+                          children: <Widget>[
+                            Column(
+                              children: <Widget>[
+                                Flexible(
+                                  child: GestureDetector(
+                                    onTap: () {
+                                      Navigator.push(
+                                        context,
+                                        MaterialPageRoute(
+                                          builder: (context) =>
+                                              Blog_detail(blog: blogs[index]),
+                                        ),
+                                      );
+                                    },
+                                    child: ClipRRect(
+                                      borderRadius: BorderRadius.circular(6.0),
+                                      child: Stack(
+                                        children: [
+                                          Image.network(
+                                            blogs[index].imageUrl,
+                                            fit: BoxFit.cover,
+                                            width: 230,
+                                            height: 180,
+                                          ),
+                                          Positioned(
+                                            right: 10,
+                                            top: 10,
+                                            child: Container(
+                                              width: 45,
+                                              height: 18,
+                                              decoration: BoxDecoration(
+                                                color: const Color(0xFF7351E6),
+                                                borderRadius:
+                                                    BorderRadius.circular(15),
+                                              ),
+                                              child: const Center(
+                                                child: Text(
+                                                  'New',
+                                                  style: TextStyle(
+                                                    color: Colors.white,
+                                                    fontSize: 12,
+                                                  ),
+                                                ),
+                                              ),
+                                            ),
+                                          ),
+                                        ],
+                                      ),
+                                    ),
+                                  ),
+                                ),
+                                SizedBox(
+                                  width: 160,
+                                  height: 84,
+                                  child: Column(
+                                    children: <Widget>[
+                                      const SizedBox(height: 5),
+                                      Text(blogs[index].title,
+                                          style: const TextStyle(
+                                              fontSize: 14,
+                                              fontWeight: FontWeight.bold)),
+                                    ],
+                                  ),
+                                )
+                              ],
+                            ),
+                          ],
+                        ),
+                      ),
+                    );
+                  },
+                ),
+              ),
+            ],
+          );
+        } else {
+          return const Text('No data');
+        }
+      },
+    );
+  }
+}
